# edcconnectoradapter-main/openapi/openapi-v1.yaml
openapi: 3.0.1
info:
  title: EDC Connector Adapter API
  description: |-
    APIs enable seamless communication with the EDC Connector, supporting a modular architecture that separates the responsibilities of asset/policy registration, contract negotiation and data consumption. This set of APIs facilitates secure, standardized, and traceable data exchange between providers and consumers in federated data ecosystems.

    ## Business Context

    The EDC Connector Adapter acts as an intermediary between provider-side and consumer-side applications and the EDC Connector. It enables:

    1. Registration and enrichment of assets and policies on the provider side.

    2. Discovery of data offers from provider connector catalog.

    3. Negotiation of usage contracts based on defined policies.

    4. Initiation and monitoring of data transfers.


    ### Key Actors

    - **Provider**: Publishes data assets, defines contract terms, and registers policies.

    - **Consumer**: Searches for offers, negotiates contracts, and consumes data.

    ### Typical Flow

    1. Retrieve Participant Identity

    2. Register and Enrich Assets and Policies

    3. Retrieve Asset Details

    4. Initiate Contract Negotiation

    5. Check Contract Status

    6. Initiate Data Transfer

    7. Monitor Transfer Status
  contact:
    name: Simpl Programme
    url: https://simpl-programme.ec.europa.eu/
    email: cnect-simpl@ec.europa.eu
  license:
...
  version: "1.0"
servers:
...
security:
- NoAuth: []
paths:
...
  /contracts:
    post:
      tags:
      - Contract
      summary: Initiate a new contract negotiation
      description: Initiates a contract negotiation between a consumer and a provider for a selected data offer. This step is required before any data transfer can occur.
      operationId: startContractNegotiation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractNegotiationRequest'
            examples:
              ContractNegotiationRequest:
                description: ContractNegotiationRequest
                value:
                  assetId: asset-123
                  contractDefinitionId: 961ead7f-d8b2-4fc6-8e94-e3efbe2127ea
                  providerEndpoint: http://endpoint.provider.com
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
                    instance: /contracts
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
                    issues:
                    - detail: Field 'assetId' must not be empty
                    instance: /contracts
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractNegotiationId'
              examples:
                contractNegotiationId:
                  description: contractNegotiationId
                  value:
                    contractNegotiationId: negotiation-abcd1234
  /connectorCatalog/assets:
    post:
      tags:
      - Contract
      summary: Returns detailed information about the asset available in the connector's catalog, useful for initiating a contract negotiation.
      description: Allows a consumer to extract detailed information about a specific asset of interest from the provider connector's catalog. This provides the preliminary information necessary to initiate the contract negotiation phase.
      operationId: searchCatalogOffers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractNegotiationRequest'
            examples:
              ContractNegotiationRequest:
                description: ContractNegotiationRequest
                value:
                  assetId: asset-123
                  contractDefinitionId: 961ead7f-d8b2-4fc6-8e94-e3efbe2127ea
                  providerEndpoint: http://endpoint.provider.com
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
                    instance: /catalog/assetDetails
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
                    issues:
                    - detail: Field 'assetId' must not be empty
                    instance: /catalog/assetDetails
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogSearchResult'
              examples:
                catalogResponse:
                  description: catalogResponse
                  value:
                    offers:
                    - assetId: asset-123-xyz
                      offerId: offer-456-abc
                      policy:
                        id: policy-789-def
                        policyConstraints:
                        - condition: PURPOSE
                          conditionOperator: EQ
                          conditionValue: data-sharing
                        type: USAGE
                      providerEndpointUrl: https://provider-edc.example.com/api/v1/dsp
                      providerParticipantId: provider-id-123456
                    - assetId: asset-789-uvw
                      offerId: offer-321-rst
                      policy:
                        id: policy-654-ghi
                        policyConstraints: []
                        type: ACCESS
                      providerEndpointUrl: https://another-provider.example.com/api/v1/dsp
                      providerParticipantId: provider-id-987654
...
  /contracts/{contractNegotiationId}:
    get:
      tags:
      - Contract
      summary: Returns the contract negotiation status
      description: Retrieves the current status of a contract negotiation. Useful for tracking the progress and outcome of the negotiation process.
      operationId: getContractNegotiationStatus
      parameters:
      - name: contractNegotiationId
        in: path
        description: The ID of the contract negotiation to retrieve the status for
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
                    instance: /contracts/{contractNegotiationId}
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
                    issues:
                    - detail: Field 'contractId' must not be empty
                    instance: /contracts/{contractNegotiationId}
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractNegotiation'
              examples:
                contractNegotiation:
                  description: contractNegotiation
                  value:
                    id: negotiation-abcd1234
                    state: CONFIRMED
                    counterPartyId: provider-edc-123
                    counterPartyAddress: https://provider-edc.example.com
                    contractAgreementId: agreement-xyz789
...
components:
  schemas:
    InternalServerErrorProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    BadRequestProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        issues:
          type: array
          items:
            $ref: '#/components/schemas/InputValidationIssue'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
    InputValidationIssue:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/InputObject'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
    ContractNegotiationRequest:
      required:
      - assetId
      - contractDefinitionId
      - providerEndpoint
      type: object
      properties:
        providerEndpoint:
          minLength: 1
          type: string
        contractDefinitionId:
          minLength: 1
          type: string
        assetId:
          minLength: 1
          type: string
    ContractNegotiationId:
      type: object
      properties:
        contractNegotiationId:
          type: string
    CatalogSearchResult:
      type: object
      properties:
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Dataset'
    Dataset:
      type: object
      properties:
        providerParticipantId:
          type: string
        providerEndpointUrl:
          type: string
        assetId:
          type: string
        offerId:
          type: string
        policy:
          $ref: '#/components/schemas/Policy'
...
    ContractNegotiation:
      type: object
      properties:
        contractAgreementId:
          type: string
        state:
          type: string
        counterPartyAddress:
          type: string
        counterPartyId:
          type: string
        errorDetail:
          type: string
        protocol:
          type: string
        type:
          type: string
        createdAt:
          type: integer
          format: int64
        '@id':
          type: string
...
  securitySchemes:
    NoAuth:
      type: apiKey
      description: Placeholder scheme used when security is not required
      name: X-No-Auth
      in: header